+++ src/pg/mod.rs
@@ -1100,7 +1100,6 @@ mod tests {
     use diesel::sql_types::Integer;
     use diesel::IntoSql;
     use futures_util::future::try_join;
-    use futures_util::try_join;
     use scoped_futures::ScopedFutureExt;
 
     #[tokio::test]
@@ -1118,7 +1117,7 @@ mod tests {
         let f1 = q1.get_result::<i32>(&mut conn);
         let f2 = q2.get_result::<i32>(&mut conn);
 
-        let (r1, r2) = try_join!(f1, f2).unwrap();
+        let (r1, r2) = try_join(f1, f2).await.unwrap();
 
         assert_eq!(r1, 1);
         assert_eq!(r2, 2);
@@ -1137,20 +1136,20 @@ mod tests {
             let f1 = diesel::select(1_i32.into_sql::<Integer>()).get_result::<i32>(&mut conn);
             let f2 = diesel::select(2_i32.into_sql::<Integer>()).get_result::<i32>(&mut conn);
 
-            try_join!(f1, f2)
+            try_join(f1, f2).await
         }
 
         async fn fn34(mut conn: &AsyncPgConnection) -> QueryResult<(i32, i32)> {
             let f3 = diesel::select(3_i32.into_sql::<Integer>()).get_result::<i32>(&mut conn);
             let f4 = diesel::select(4_i32.into_sql::<Integer>()).get_result::<i32>(&mut conn);
 
-            try_join!(f3, f4)
+            try_join(f3, f4).await
         }
 
         let f12 = fn12(&conn);
         let f34 = fn34(&conn);
 
-        let ((r1, r2), (r3, r4)) = try_join!(f12, f34).unwrap();
+        let ((r1, r2), (r3, r4)) = try_join(f12, f34).await.unwrap();
 
         assert_eq!(r1, 1);
         assert_eq!(r2, 2);
@@ -1189,7 +1188,7 @@ mod tests {
             let f5 = diesel::select(5_i32.into_sql::<Integer>()).get_result::<i32>(&mut conn);
             let f6 = diesel::select(6_i32.into_sql::<Integer>()).get_result::<i32>(&mut conn);
 
-            try_join!(f5.boxed(), f6.boxed())
+            try_join(f5.boxed(), f6.boxed()).await
         }
 
         conn.transaction(|conn| {
@@ -1198,7 +1197,8 @@ mod tests {
                 let f34 = fn34(conn);
                 let f56 = fn56(conn);
 
-                let ((r1, r2), (r3, r4), (r5, r6)) = try_join!(f12, f34, f56).unwrap();
+                let ((r1, r2), ((r3, r4), (r5, r6))) =
+                    try_join(f12, try_join(f34, f56)).await.unwrap();
 
                 assert_eq!(r1, 1);
                 assert_eq!(r2, 2);